feat: Se agrega Readme con explicación para integración en app.
