+++ src/common/interfaces/deleteProperty/deleteProterty.ibterface.ts
@@ -1,6 +1,3 @@
-// src/common/interfaces/deleteProperty/deleteProperty.interface.ts
-
 export interface IDeleteProperty {
-    id: number;
-  }
-  
+  id: string; // UUID como string
+}
