+++ src/dtos/deletePropertyDTO/deleteProperty.dto.ts
@@ -1,0 +1,7 @@
+import { IsString, IsNotEmpty } from 'class-validator';
+
+export class DeletePropertyDto {
+  @IsString()
+  @IsNotEmpty()
+  id: string; // UUID como string
+}
